requirements.txt
streamlit
chromadb
sentence-transformers
transformers
torch
pdfplumber
pillow
pytesseract
numpy
scikit-learn
reportlab

